Letak dari getPOLength sebelumnya salah, seharusnya berada pada model PurchaseOrder malah sebelumnya berada pada controller
